// src/app/api/models/Level.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Cursus } from './Cursus';
export type Level = {
    id?: string;
    name: string;
    icon?: string | null;
    color?: string | null;
    cursus?: Array<Cursus> | null;
};